ci: add constructor test

+++ src/main/java/dev/nyanchuk/fizzbuzz/App.java
@@ -5,7 +5,7 @@ public final class App {
     }
 
     public static void main(String[] args) {
-        for (int num = 1; num <= 100; num++) {
+        for (byte num = 1; num <= 100; num++) {
             System.out.println(Fizzbuzz.FizzBuzzEvaluator1(num));
         }
     }
